Move libObjectModel code
