+++ src/CMakeLists.txt
@@ -271,6 +271,7 @@ if(BUILD_UI)
 	add_subdirectory(ui)		#all antares ui libs + antares simulator
 endif()
 
+add_subdirectory(study)	    #antares study model
 add_subdirectory(solver)	#antares solver and all associated libs
 add_subdirectory(analyzer)	#antares analyser
