+++ src/solver/CMakeLists.txt
@@ -12,7 +12,6 @@ add_subdirectory(constraints-builder)
 add_subdirectory(expressions)
 add_subdirectory(hydro)
 add_subdirectory(infeasible-problem-analysis)
-add_subdirectory(libModelObject)
 add_subdirectory(lps)
 add_subdirectory(misc)
 add_subdirectory(modelConverter)
