+++ src/study/CMakeLists.txt
@@ -1,0 +1,7 @@
+OMESSAGE("Antares Study")
+
+add_subdirectory(system-model)
+
+install(DIRECTORY include/antares
+        DESTINATION "include"
+)
