+++ src/study/system-model/include/antares/study/system-model/variable.h
@@ -19,18 +19,52 @@
 ** along with Antares_Simulator. If not, see <https://opensource.org/license/mpl-2-0/>.
 */
 #pragma once
-#include <iostream>
+
 #include <string>
 
-namespace Antares::Solver::ObjectModel
+#include "expression.h"
+#include "valueType.h"
+
+namespace Antares::Study::SystemModel
 {
 
-/// Type of value held by variables or parameters
-enum class ValueType
+/// A decision variable of the model
+class Variable
 {
-    FLOAT,
-    INTEGER,
-    BOOL
+public:
+    Variable(std::string id, Expression lower_bound, Expression upper_bound, ValueType type):
+        id_(std::move(id)),
+        type_(type),
+        lowerBound_(lower_bound),
+        upperBound_(upper_bound)
+    {
+    }
+
+    const std::string& Id() const
+    {
+        return id_;
+    }
+
+    ValueType Type() const
+    {
+        return type_;
+    }
+
+    Expression LowerBound() const
+    {
+        return lowerBound_;
+    }
+
+    Expression UpperBound() const
+    {
+        return upperBound_;
+    }
+
+private:
+    std::string id_;
+    ValueType type_;
+    Expression lowerBound_;
+    Expression upperBound_;
 };
 
-} // namespace Antares::Solver::ObjectModel
+} // namespace Antares::Study::SystemModel
