+++ src/tests/src/solver/modelParser/testModelTranslator.cpp
@@ -25,19 +25,20 @@
 
 #include <boost/test/unit_test.hpp>
 
-#include "antares/solver/libObjectModel/library.h"
+#include "antares/study/system-model/library.h"
 #include "antares/solver/modelConverter/modelConverter.h"
 #include "antares/solver/modelParser/Library.h"
 
 #include "enum_operators.h"
 
 using namespace Antares::Solver;
+using namespace Antares::Study;
 
 // Test empty library
 BOOST_AUTO_TEST_CASE(Empty_library_is_valid)
 {
     ModelParser::Library library;
-    ObjectModel::Library lib = ModelConverter::convert(library);
+    SystemModel::Library lib = ModelConverter::convert(library);
     BOOST_CHECK(lib.Id().empty());
     BOOST_CHECK(lib.Description().empty());
     BOOST_CHECK(lib.PortTypes().empty());
@@ -50,7 +51,7 @@ BOOST_AUTO_TEST_CASE(library_id_description_properly_translated)
     ModelParser::Library library;
     library.id = "test_id";
     library.description = "test_description";
-    ObjectModel::Library lib = ModelConverter::convert(library);
+    SystemModel::Library lib = ModelConverter::convert(library);
     BOOST_CHECK_EQUAL(lib.Id(), "test_id");
     BOOST_CHECK_EQUAL(lib.Description(), "test_description");
 }
@@ -62,7 +63,7 @@ BOOST_AUTO_TEST_CASE(port_type_with_empty_fileds_properly_translated)
     ModelParser::PortType portType1{"port1", "flow port", {}};
     ModelParser::PortType portType2{"port2", "impedance port", {}};
     library.port_types = {portType1, portType2};
-    ObjectModel::Library lib = ModelConverter::convert(library);
+    SystemModel::Library lib = ModelConverter::convert(library);
     BOOST_REQUIRE_EQUAL(lib.PortTypes().size(), 2);
     BOOST_CHECK_EQUAL(lib.PortTypes().at("port1").Id(), "port1");
     BOOST_CHECK_EQUAL(lib.PortTypes().at("port1").Description(), "flow port");
@@ -86,7 +87,7 @@ BOOST_AUTO_TEST_CASE(portType_with_fields_properly_translated)
     ModelParser::PortType portType1{"port1", "flow port", {"field1", "field2"}};
     ModelParser::PortType portType2{"port2", "impedance port", {"field3", "field4"}};
     library.port_types = {portType1, portType2};
-    ObjectModel::Library lib = ModelConverter::convert(library);
+    SystemModel::Library lib = ModelConverter::convert(library);
     BOOST_REQUIRE_EQUAL(lib.PortTypes().at("port1").Fields().size(), 2);
     BOOST_CHECK_EQUAL(lib.PortTypes().at("port1").Fields()[0].Id(), "field1");
     BOOST_CHECK_EQUAL(lib.PortTypes().at("port1").Fields()[1].Id(), "field2");
@@ -108,7 +109,7 @@ BOOST_AUTO_TEST_CASE(empty_model_properly_translated)
                               .constraints = {},
                               .objective = "objectives"};
     library.models = {model1};
-    ObjectModel::Library lib = ModelConverter::convert(library);
+    SystemModel::Library lib = ModelConverter::convert(library);
     BOOST_REQUIRE_EQUAL(lib.Models().size(), 1);
     BOOST_CHECK_EQUAL(lib.Models().at("model1").Id(), "model1");
     BOOST_CHECK_EQUAL(lib.Models().at("model1").Objective().Value(), "objectives");
@@ -127,7 +128,7 @@ BOOST_AUTO_TEST_CASE(model_parameters_properly_translated)
                               .constraints{},
                               .objective = "objectives"};
     library.models = {model1};
-    ObjectModel::Library lib = ModelConverter::convert(library);
+    SystemModel::Library lib = ModelConverter::convert(library);
     auto& model = lib.Models().at("model1");
     BOOST_REQUIRE_EQUAL(model.Parameters().size(), 2);
     auto& parameter1 = model.Parameters().at("param1");
@@ -135,11 +136,11 @@ BOOST_AUTO_TEST_CASE(model_parameters_properly_translated)
     BOOST_CHECK_EQUAL(parameter1.Id(), "param1");
     BOOST_CHECK(parameter1.isTimeDependent());
     BOOST_CHECK(!parameter1.isScenarioDependent());
-    BOOST_CHECK_EQUAL(parameter1.Type(), ObjectModel::ValueType::FLOAT);
+    BOOST_CHECK_EQUAL(parameter1.Type(), SystemModel::ValueType::FLOAT);
     BOOST_CHECK_EQUAL(parameter2.Id(), "param2");
     BOOST_CHECK(!parameter2.isTimeDependent());
     BOOST_CHECK(!parameter2.isScenarioDependent());
-    BOOST_CHECK_EQUAL(parameter2.Type(), ObjectModel::ValueType::FLOAT);
+    BOOST_CHECK_EQUAL(parameter2.Type(), SystemModel::ValueType::FLOAT);
 }
 
 // Test library with models and variables
@@ -157,7 +158,7 @@ BOOST_AUTO_TEST_CASE(model_variables_properly_translated)
       .constraints = {},
       .objective = "objectives"};
     library.models = {model1};
-    ObjectModel::Library lib = ModelConverter::convert(library);
+    SystemModel::Library lib = ModelConverter::convert(library);
     auto& model = lib.Models().at("model1");
     BOOST_REQUIRE_EQUAL(model.Variables().size(), 2);
     auto& variable1 = model.Variables().at("var1");
@@ -165,11 +166,11 @@ BOOST_AUTO_TEST_CASE(model_variables_properly_translated)
     BOOST_CHECK_EQUAL(variable1.Id(), "var1");
     BOOST_CHECK_EQUAL(variable1.LowerBound().Value(), "7");
     BOOST_CHECK_EQUAL(variable1.UpperBound().Value(), "pmax");
-    BOOST_CHECK_EQUAL(variable1.Type(), ObjectModel::ValueType::BOOL);
+    BOOST_CHECK_EQUAL(variable1.Type(), SystemModel::ValueType::BOOL);
     BOOST_CHECK_EQUAL(variable2.Id(), "var2");
     BOOST_CHECK_EQUAL(variable2.LowerBound().Value(), "99999999.9999999");
     BOOST_CHECK_EQUAL(variable2.UpperBound().Value(), "vcost");
-    BOOST_CHECK_EQUAL(variable2.Type(), ObjectModel::ValueType::INTEGER);
+    BOOST_CHECK_EQUAL(variable2.Type(), SystemModel::ValueType::INTEGER);
 }
 
 // Test library with models and ports
@@ -185,7 +186,7 @@ BOOST_AUTO_TEST_CASE(model_ports_properly_translated, *boost::unit_test::disable
                               .constraints = {},
                               .objective = "objectives"};
     library.models = {model1};
-    ObjectModel::Library lib = ModelConverter::convert(library);
+    SystemModel::Library lib = ModelConverter::convert(library);
     auto& model = lib.Models().at("model1");
     // BOOST_REQUIRE_EQUAL(model.Ports().size(), 2);
     // auto& port1 = model.Ports().at("port1");
@@ -210,7 +211,7 @@ BOOST_AUTO_TEST_CASE(model_constraints_properly_translated)
                                               {"constraint2", "expression2"}},
                               .objective = "objectives"};
     library.models = {model1};
-    ObjectModel::Library lib = ModelConverter::convert(library);
+    SystemModel::Library lib = ModelConverter::convert(library);
     auto& model = lib.Models().at("model1");
     BOOST_REQUIRE_EQUAL(model.getConstraints().size(), 2);
     auto& constraint1 = model.getConstraints().at("constraint1");
@@ -245,7 +246,7 @@ BOOST_AUTO_TEST_CASE(multiple_models_properly_translated)
       .constraints = {},
       .objective = "objectives"};
     library.models = {model1, model2};
-    ObjectModel::Library lib = ModelConverter::convert(library);
+    SystemModel::Library lib = ModelConverter::convert(library);
     BOOST_REQUIRE_EQUAL(lib.Models().size(), 2);
     auto& modelo1 = lib.Models().at("model1");
     BOOST_REQUIRE_EQUAL(modelo1.Parameters().size(), 2);
